rightscale_clustering.sh
This script takes a series of inputs and automatically creates a cluster.  It is designed to run in parallel on an arbitrary number of nodes as well as arbitrary service configuration.

There's a bunch of extra code in here for handling different versions as well as some rather ugly retry loops and error handling.